requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
Django==2.2.15
django-celery-beat==2.0.0
django-timezone-field==4.0
djangorestframework==3.11.0
flake8==3.8.3
importlib-metadata==1.7.0
kombu==4.6.11
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==1.2.0
